requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-dotenv==0.19.0
rasa==3.1.0
SpeechRecognition==3.8.1
pyttsx3==2.90
pyaudio==0.2.11